Make build tests wait 20ms after writing to files.
This lets the tests pass on file systems with coarse
file modification timing.

+++ src/ebb/src/stdext/src/stdext/string_view.h
@@ -23,7 +23,7 @@ struct hash<stdext::string_view>
     // implementation since we call a 32-bit hashing function on a (possibly)
     // 64 bit value.  Oh well!
     std::size_t result = 0;
-    MurmurHash3_x86_32(key.data(), key.size(), 0, &result);
+    MurmurHash3_x86_32(key.data(), static_cast<int>(key.size()), 0, &result);
     return result;
   }
 };
